Extract bewits in RequestBuilder
Rather than do string manipulation in the Bewit implementation, allow
extracting a Bewit from request metadata while building the Request
object.  The result is a Request with a properly-formed `path` (that is,
without the `bewit=..`) and a Bewit for the caller to validate.

This is a breaking API change to the previous Bewit implementation,
since it drops the `from_path` method, but that implementation has not
been released.